requirements.txt
Django>=5.2
python-decouple
social-auth-app-django
django-crispy-forms
crispy-bootstrap5
gunicorn
whitenoise
dj-database-url
psycopg2-binary
